# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Clima",
        "description": "API para consultar datos del clima y gestionar el historial de búsquedas.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/searches": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Obtener historial de búsquedas",
                "description": "Devuelve el historial de búsquedas realizadas por todos los usuarios. Requiere autenticación y rol admin.",
                "operationId": "f03a75c9bde0ab48d9b0ebcdb47e880b",
                "responses": {
                    "200": {
                        "description": "Historial de búsquedas obtenido correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "all_search_history": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "city": {
                                                        "type": "string",
                                                        "example": "Paris"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-04-03T19:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acceso no autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registrar un nuevo usuario",
                "description": "Crea un nuevo usuario y devuelve el token de autenticación.",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "description": "Datos del usuario a registrar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Registro exitoso."
                                        },
                                        "user": {
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Autenticar usuario",
                "description": "Inicia sesión y devuelve un token de autenticación.",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "description": "Credenciales del usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Inicio de sesión exitoso."
                                        },
                                        "user": {
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Cerrar sesión",
                "description": "Revoca el token de autenticación actual.",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Cierre de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sesión cerrada correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/weather": {
            "get": {
                "tags": [
                    "Weather"
                ],
                "summary": "Obtener datos del clima de una ciudad",
                "description": "Consulta la API de WeatherAPI para obtener datos del clima (temperatura, humedad, viento, etc.). Se requiere autenticación mediante Bearer token.",
                "operationId": "fa8cc72818e1e45b2ff1c61a16e1c876",
                "parameters": [
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Nombre de la ciudad a consultar",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Paris"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del clima obtenidos correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "city": {
                                            "type": "string",
                                            "example": "Paris"
                                        },
                                        "region": {
                                            "type": "string",
                                            "example": "Île-de-France"
                                        },
                                        "country": {
                                            "type": "string",
                                            "example": "France"
                                        },
                                        "local_time": {
                                            "type": "string",
                                            "example": "2025-04-03 19:00"
                                        },
                                        "temperature": {
                                            "type": "number",
                                            "example": 18
                                        },
                                        "humidity": {
                                            "type": "number",
                                            "example": 50
                                        },
                                        "wind_kph": {
                                            "type": "number",
                                            "example": 10
                                        },
                                        "condition": {
                                            "type": "string",
                                            "example": "Clear"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ciudad no encontrada o sin datos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/history": {
            "get": {
                "tags": [
                    "Weather"
                ],
                "summary": "Obtener historial de búsquedas",
                "description": "Retorna el historial de búsquedas realizadas por el usuario autenticado.",
                "operationId": "fe759671b48d299e7d095658474a179e",
                "responses": {
                    "200": {
                        "description": "Historial obtenido correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "history": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "city": {
                                                        "type": "string",
                                                        "example": "Paris"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-04-03T19:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/favorites": {
            "get": {
                "tags": [
                    "Weather"
                ],
                "summary": "Listar ciudades favoritas",
                "description": "Retorna la lista de ciudades marcadas como favoritas por el usuario autenticado.",
                "operationId": "1a73e96826ddc6a640454b6927f6fd58",
                "responses": {
                    "200": {
                        "description": "Favoritos obtenidos correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "favorites": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "city": {
                                                        "type": "string",
                                                        "example": "Paris"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-04-03T19:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Weather"
                ],
                "summary": "Agregar ciudad a favoritos",
                "description": "Marca una ciudad como favorita para el usuario autenticado.",
                "operationId": "7d8e87349e05042130c4320a20cd7a52",
                "requestBody": {
                    "description": "Ciudad a agregar como favorita",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "city"
                                ],
                                "properties": {
                                    "city": {
                                        "type": "string",
                                        "example": "Paris"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ciudad agregada a favoritos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ciudad agregada a favoritos."
                                        },
                                        "favorite": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "SearchHistory": {
                "title": "Historial de Búsqueda",
                "required": [
                    "id",
                    "city",
                    "created_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "city": {
                        "type": "string",
                        "example": "Paris"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-03T19:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Ingresa el token en formato (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Admin",
            "description": "Admin"
        },
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Weather",
            "description": "Weather"
        }
    ],
    "security": [
        {
            "sanctum": []
        }
    ]
}